Refactor of Neo-Lux codebase
