+++ neo-sender/Program.cs
@@ -14,8 +14,6 @@ namespace neo_sender
                 return;
             }
 
-            var net = (NeoAPI.Net) Enum.Parse(typeof(NeoAPI.Net), args[0], true);
-
             var keyStr = args[1];
             //fc1fa7c0d83426486373d9ce6eaca8adb506fc4ca25e89887c8eb5567f317a53"
             var outputAddress = args[2];
@@ -28,7 +26,17 @@ namespace neo_sender
 
             Console.WriteLine($"Sending {amount} {symbol} from {myKey.address} to {outputAddress}");
 
-            var result = NeoAPI.SendAsset(net, outputAddress, symbol, amount, myKey);
+            var net = args[0].ToLowerInvariant();
+            NeoAPI api;
+
+            switch (net)
+            {
+                case "main": api = NeoRPC.ForMainNet(); break;
+                case "test": api = NeoRPC.ForTestNet(); break;
+                default: api = new NeoRPC(net); break;
+            }
+
+            var result = api.SendAsset(outputAddress, symbol, amount, myKey);
             Console.WriteLine(result);
         }
     }
